setup open and closed versions
