+++ src/utils/constants.ts
@@ -13,3 +13,5 @@ export const COOKIE_POLICY_PAGE = '/cookie-policy'
 export const CONTENT_DIR = 'public/content'
 
 export const MARKDOWN_CONTENT_MAX_WIDTH = 800
+
+export const SCHOLAR_PROGRAM_OPEN = true
